requirements.txt
numba==0.54.0
numpy==1.20.3
scipy==1.7.1
matplotlib==3.4.3
pyglet==1.5.0
catkin_pkg==0.4.23
cvxpy==1.1.15
docker_py==1.10.6
ffmpeg
gym==0.21.0
jsonlines==2.0.0
osqp==0.6.2.post0
Pillow==8.4.0
PyYAML==6.0
pyzmq==22.3.0
Shapely==1.7.1
sphinx_rtd_theme==1.0.0
stable_baselines3==1.2.0
tf==1.0.0
wandb==0.12.7
casadi
